# OPENAPI.yaml
...
openapi: 3.1.0
info:
  title: netflat API
  description: |
    REST API cho ứng dụng xem video tự sản xuất theo trải nghiệm Netflix.
    
    ## Authentication
    API sử dụng JWT Bearer token. Access token có thời hạn 15 phút, refresh token 7 ngày.
    
    ## Error Format
    Tất cả lỗi trả về theo format:
    ```json
    {
      "error": {
        "code": "ERROR_CODE",
        "message": "Human readable message",
        "details": null,
        "requestId": "uuid"
      }
    }
    ```
  version: 1.0.0
  contact:
    name: netflat Team
    email: dev@netflat.local
servers:
  - url: http://localhost:3000/api
    description: Local development
  - url: https://api.netflat.example.com/api
    description: Staging (placeholder)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 5
        hasNext:
          type: boolean
          example: true
        hasPrev:
          type: boolean
          example: false
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - requestId
          properties:
            code:
              type: string
              example: VALIDATION_FAILED
            message:
              type: string
              example: Validation failed
            details:
              type: object
              nullable: true
              additionalProperties: true
            requestId:
              type: string
              format: uuid
              example: 550e8400-e29b-41d4-a716-446655440000

    # ─────────────────────────────────────────────────────────────
    # Auth
    # ─────────────────────────────────────────────────────────────
...
    Movie:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        title:
          type: string
          example: "Sample Action Movie"
        description:
          type: string
          nullable: true
          example: "An exciting action-packed adventure."
        posterUrl:
          type: string
          nullable: true
          example: "https://storage.example.com/posters/movie-id/poster.jpg"
        backdropUrl:
          type: string
          nullable: true
          example: "https://storage.example.com/posters/movie-id/backdrop.jpg"
        durationSeconds:
          type: integer
          nullable: true
          example: 7200
        releaseYear:
          type: integer
          nullable: true
          example: 2024
        movieStatus:
          type: string
          enum: [draft, published]
          example: published
        encodeStatus:
          type: string
          enum: [pending, processing, ready, failed]
          example: ready
        genres:
          type: array
          items:
            $ref: '#/components/schemas/Genre'
        createdAt:
          type: string
          format: date-time
          example: "2026-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-01-01T00:00:00Z"
    MovieListItem:
      type: object
      description: Simplified movie for list/rail display
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        posterUrl:
          type: string
          nullable: true
        durationSeconds:
          type: integer
          nullable: true
        releaseYear:
          type: integer
          nullable: true
        genres:
          type: array
          items:
            $ref: '#/components/schemas/Genre'
    MovieListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/MovieListItem'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    MovieDetailResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Movie'
    StreamResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            playbackUrl:
              type: string
              description: |
                HLS master playlist URL. In public mode this is a public URL; in signed mode it's a presigned URL.
              example: "https://storage.example.com/hls/movie-id/master.m3u8?token=xxx&expires=xxx"
            qualityOptions:
              type: array
              description: Optional variant playlists for manual quality selection
              items:
                type: object
                properties:
                  name:
                    type: string
                    example: "720p"
                  url:
                    type: string
                    example: "https://storage.example.com/hls/movie-id/v2/prog_index.m3u8"
            expiresAt:
              type: string
              format: date-time
              nullable: true
              description: May be null in public mode
              example: "2026-01-01T01:00:00Z"
...
    CreateMovieRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          maxLength: 500
          example: "My New Movie"
        description:
          type: string
          nullable: true
          example: "A great movie about..."
        genreIds:
          type: array
          items:
            type: string
            format: uuid
          example: ["550e8400-e29b-41d4-a716-446655440001", "550e8400-e29b-41d4-a716-446655440002"]
        releaseYear:
          type: integer
          nullable: true
          example: 2024
    UpdateMovieRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 500
          example: "Updated Title"
        description:
          type: string
          nullable: true
        genreIds:
          type: array
          items:
            type: string
            format: uuid
        releaseYear:
          type: integer
          nullable: true
        durationSeconds:
          type: integer
          nullable: true
    PublishRequest:
      type: object
      required:
        - published
      properties:
        published:
          type: boolean
          example: true

    # ─────────────────────────────────────────────────────────────
    # Genres
    # ─────────────────────────────────────────────────────────────
    Genre:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440001
        name:
          type: string
          example: Action
        slug:
          type: string
          example: action
...
    PresignedUrlResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            uploadUrl:
              type: string
              description: Presigned PUT URL
              example: "https://storage.example.com/originals/movie-id/original.mp4?X-Amz-..."
            objectKey:
              type: string
              description: Object key for upload-complete callback
              example: "originals/550e8400-e29b-41d4-a716-446655440000/original.mp4"
            expiresAt:
              type: string
              format: date-time
              example: "2026-01-01T00:30:00Z"
    UploadCompleteRequest:
      type: object
      required:
        - objectKey
      properties:
        objectKey:
          type: string
          example: "originals/550e8400-e29b-41d4-a716-446655440000/original.mp4"
        fileType:
          type: string
          enum: [video, thumbnail]
          default: video
          example: video
    UploadCompleteResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            movieId:
              type: string
              format: uuid
            encodeStatus:
              type: string
              enum: [pending, processing, ready, failed]
              example: pending
            jobId:
              type: string
              description: BullMQ job ID (for video uploads)
              example: "encode-123"
              nullable: true
            status:
              type: string
              description: Upload/encode enqueue status
              example: queued
              nullable: true
            posterUrl:
              type: string
              description: Poster URL when fileType=thumbnail
              nullable: true
...
paths:
...
  /movies:
    get:
      tags: [Movies]
      summary: List movies
      description: |
        Get paginated list of published movies. 
        Supports filtering by genre and search by keyword.
      operationId: listMovies
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: genreId
          in: query
          description: Filter by genre ID
          schema:
            type: string
            format: uuid
        - name: q
          in: query
          description: Search keyword (title/description)
          schema:
            type: string
        - name: continueWatching
          in: query
          description: Get user's continue watching list
          schema:
            type: boolean
        - name: sort
          in: query
          description: Sort field
          schema:
            type: string
            enum: [createdAt, title, releaseYear]
            default: createdAt
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of movies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieListResponse'
              example:
                data:
                  - id: "550e8400-e29b-41d4-a716-446655440000"
                    title: "Sample Action Movie"
                    posterUrl: "https://storage.example.com/posters/movie-id/poster.jpg"
                    durationSeconds: 7200
                    releaseYear: 2024
                    genres:
                      - id: "550e8400-e29b-41d4-a716-446655440001"
                        name: "Action"
                        slug: "action"
                meta:
                  page: 1
                  limit: 20
                  total: 100
                  totalPages: 5
                  hasNext: true
                  hasPrev: false
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags: [Admin]
      summary: Create movie (Admin)
      description: Create a new movie draft. Requires admin role.
      operationId: createMovie
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMovieRequest'
      responses:
        '201':
          description: Movie created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieDetailResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /upload/complete/{movieId}:
    post:
      tags: [Upload]
      summary: Upload complete callback (Admin, deprecated)
      description: |
        Deprecated alias. Use `POST /movies/{id}/upload-complete` instead.
        For video uploads, this triggers the encode job.
      operationId: uploadCompleteLegacy
      deprecated: true
      security:
        - bearerAuth: []
      parameters:
        - name: movieId
          in: path
          required: true
          description: Movie ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadCompleteRequest'
      responses:
        '200':
          description: Upload processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadCompleteResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: FORBIDDEN_ADMIN_ONLY
                  message: "This action requires admin role"
                  details: null
                  requestId: "550e8400-e29b-41d4-a716-446655440000"
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /movies/{id}:
    get:
      tags: [Movies]
      summary: Get movie detail
      description: Get full movie details by ID
      operationId: getMovie
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Movie detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieDetailResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Movie not available for viewing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: MOVIE_NOT_AVAILABLE
                  message: "Movie is not available"
                  details: null
                  requestId: "550e8400-e29b-41d4-a716-446655440000"
        '404':
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: MOVIE_NOT_FOUND
                  message: "Movie not found"
                  details: null
                  requestId: "550e8400-e29b-41d4-a716-446655440000"

    put:
      tags: [Admin]
      summary: Update movie (Admin)
      description: Update movie details. Requires admin role.
      operationId: updateMovie
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMovieRequest'
      responses:
        '200':
          description: Movie updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieDetailResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags: [Admin]
      summary: Delete movie (Admin)
      description: Delete a movie. Requires admin role.
      operationId: deleteMovie
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Movie deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: "Movie deleted successfully"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /movies/{id}/stream:
    get:
      tags: [Movies]
      summary: Get stream URL
      description: |
        Get signed HLS playback URL for the movie.
        Only works for published + ready movies.
      operationId: getStreamUrl
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Stream URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Movie not found or not ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /movies/{id}/publish:
    patch:
      tags: [Admin]
      summary: Publish/Unpublish movie (Admin)
      description: Toggle movie publish status. Requires admin role.
      operationId: publishMovie
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishRequest'
      responses:
        '200':
          description: Publish status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieDetailResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /movies/{id}/upload-complete:
    post:
      tags: [Upload]
      summary: Upload complete callback (Admin)
      description: |
        Canonical endpoint for upload completion.
        For video uploads, this triggers the encode job.
      operationId: uploadComplete
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Movie ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadCompleteRequest'
      responses:
        '200':
          description: Upload processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadCompleteResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ─────────────────────────────────────────────────────────────
  # Genres
  # ─────────────────────────────────────────────────────────────
...
  /upload/presigned-url:
    get:
      tags: [Upload]
      summary: Get presigned upload URL (Admin)
      description: |
        Get a presigned URL to upload a file directly to storage.
        Requires admin role.
      operationId: getPresignedUrl
      security:
        - bearerAuth: []
      parameters:
        - name: movieId
          in: query
          required: true
          description: Movie ID to upload for
          schema:
            type: string
            format: uuid
        - name: fileName
          in: query
          required: true
          schema:
            type: string
            example: "video.mp4"
        - name: contentType
          in: query
          required: true
          schema:
            type: string
            example: "video/mp4"
        - name: sizeBytes
          in: query
          required: true
          schema:
            type: integer
            format: int64
            example: 104857600
        - name: fileType
          in: query
          schema:
            type: string
            enum: [video, thumbnail]
            default: video
      responses:
        '200':
          description: Presigned URL generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresignedUrlResponse'
        '400':
          description: Invalid file type or size
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ─────────────────────────────────────────────────────────
  # History
  # ─────────────────────────────────────────────────────────
...